+++ packages/ui/src/components/property-editor/editor-items/NumberInputItem.jsx
@@ -1,6 +1,6 @@
 import { InputEditor } from "../../input-editor";
 
-export function NumberInputItem({ value, item, style, onChange }) {
+export function NumberInputItem({ value, item, style, onChange, onInput }) {
   const { min = 0, max = 100, step = 1 } = item;
   return (
     <InputEditor
@@ -12,7 +12,17 @@ export function NumberInputItem({ value, item, style, onChange }) {
       width="100%"
       style={style}
       onInput={(e) => {
-        onChange && onChange(Number(e.target.value), item);
+        // console.log(JSON.stringify(e.target.value));
+
+        // trigger all events
+        onInput && onInput(e.target.value, item, e);
+
+        if (e.target.value === "" && e.target.value === "-") {
+          // noop
+        } else {
+          // trigger only onChange event for valid number value
+          onChange && onChange(e.target.value, item, e);
+        }
       }}
     />
   );