+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @elf-framework/design-system
 
+## 0.0.252
+
+### Patch Changes
+
+- fixed number input value format
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.252
+
+## 0.0.251
+
+### Patch Changes
+
+- modify tools item selected function
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.251
+
+## 0.0.250
+
+### Patch Changes
+
+- modify tools selected function
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.250
+
 ## 0.0.244
 
 ### Patch Changes