+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @elf-framework/design-tokens
 
+## 0.0.252
+
+### Patch Changes
+
+- fixed number input value format
+
+## 0.0.251
+
+### Patch Changes
+
+- modify tools item selected function
+
+## 0.0.250
+
+### Patch Changes
+
+- modify tools selected function
+
+## 0.0.248
+
+### Patch Changes
+
+- modify Animation method
+
 ## 0.0.244
 
 ### Patch Changes